# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 報修管理系統 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /repairs:
    get:
      operationId: listRepairs
      tags: [repairs]
      summary: List all repair records
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed]
        - name: category
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high]
        - name: month
          in: query
          description: Filter by month (YYYY-MM format)
          schema:
            type: string
      responses:
        "200":
          description: List of repair records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Repair"
    post:
      operationId: createRepair
      tags: [repairs]
      summary: Create a repair record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepairInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repair"
        "400":
          description: Bad request
  /repairs/{id}:
    get:
      operationId: getRepair
      tags: [repairs]
      summary: Get a repair record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Repair record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repair"
        "404":
          description: Not found
    patch:
      operationId: updateRepair
      tags: [repairs]
      summary: Update a repair record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepairUpdate"
      responses:
        "200":
          description: Updated repair record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repair"
        "404":
          description: Not found
    delete:
      operationId: deleteRepair
      tags: [repairs]
      summary: Delete a repair record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /stats/monthly:
    get:
      operationId: getMonthlyStats
      tags: [stats]
      summary: Get monthly statistics
      parameters:
        - name: month
          in: query
          description: Month in YYYY-MM format (defaults to current month)
          schema:
            type: string
      responses:
        "200":
          description: Monthly statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyStats"
...
  /repairs/{id}/custom-values:
    put:
      operationId: setRepairCustomValues
      tags: [repairs]
      summary: Set custom field values for a repair
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomValuesPayload"
      responses:
        "200":
          description: Updated custom values
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomValue"
        "404":
          description: Repair not found
...
components:
  schemas:
...
    Repair:
      type: object
      required:
        - id
        - title
        - location
        - category
        - status
        - priority
        - reportedBy
        - reportedAt
      properties:
        id:
          type: integer
        title:
          type: string
        location:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed]
        priority:
          type: string
          enum: [low, medium, high]
        description:
          type: ["string", "null"]
        reportedBy:
          type: string
        assignedTo:
          type: ["string", "null"]
        reportedAt:
          type: string
          format: date-time
        resolvedAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
    RepairInput:
      type: object
      required:
        - title
        - location
        - category
        - priority
        - reportedBy
      properties:
        title:
          type: string
          minLength: 1
        location:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        priority:
          type: string
          enum: [low, medium, high]
        description:
          type: string
        reportedBy:
          type: string
          minLength: 1
        assignedTo:
          type: string
        notes:
          type: string
    RepairUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        location:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed]
        priority:
          type: string
          enum: [low, medium, high]
        description:
          type: string
        reportedBy:
          type: string
          minLength: 1
        assignedTo:
          type: string
        resolvedAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: string
    MonthlyStats:
      type: object
      required:
        - month
        - total
        - byStatus
        - byCategory
        - byPriority
      properties:
        month:
          type: string
        total:
          type: integer
        byStatus:
          type: object
          properties:
            pending:
              type: integer
            in_progress:
              type: integer
            completed:
              type: integer
          required: [pending, in_progress, completed]
        byCategory:
          type: object
          additionalProperties:
            type: integer
        byPriority:
          type: object
          properties:
            low:
              type: integer
            medium:
              type: integer
            high:
              type: integer
          required: [low, medium, high]
        avgResolutionDays:
          type: ["number", "null"]
...
    CustomValue:
      type: object
      required: [id, repairId, fieldId, value]
      properties:
        id:
          type: integer
        repairId:
          type: integer
        fieldId:
          type: integer
        value:
          type: string
    CustomValueInput:
      type: object
      required: [fieldId, value]
      properties:
        fieldId:
          type: integer
        value:
          type: string
    CustomValuesPayload:
      type: object
      required: [values]
      properties:
        values:
          type: array
          items:
            $ref: "#/components/schemas/CustomValueInput"
...