# lib/api-zod/src/generated/types/monthlyStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 報修管理系統 API
 * OpenAPI spec version: 0.1.0
 */
import type { MonthlyStatsByCategory } from "./monthlyStatsByCategory";
import type { MonthlyStatsByPriority } from "./monthlyStatsByPriority";
import type { MonthlyStatsByStatus } from "./monthlyStatsByStatus";

export interface MonthlyStats {
  month: string;
  total: number;
  byStatus: MonthlyStatsByStatus;
  byCategory: MonthlyStatsByCategory;
  byPriority: MonthlyStatsByPriority;
  /** @nullable */
  avgResolutionDays?: number | null;
}