# lib/api-zod/src/generated/types/repair.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 報修管理系統 API
 * OpenAPI spec version: 0.1.0
 */
import type { RepairPriority } from "./repairPriority";
import type { RepairStatus } from "./repairStatus";

export interface Repair {
  id: number;
  title: string;
  location: string;
  category: string;
  status: RepairStatus;
  priority: RepairPriority;
  /** @nullable */
  description?: string | null;
  reportedBy: string;
  /** @nullable */
  assignedTo?: string | null;
  reportedAt: Date;
  /** @nullable */
  resolvedAt?: Date | null;
  /** @nullable */
  notes?: string | null;
}