# lib/api-zod/src/generated/types/repairUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 報修管理系統 API
 * OpenAPI spec version: 0.1.0
 */
import type { RepairUpdatePriority } from "./repairUpdatePriority";
import type { RepairUpdateStatus } from "./repairUpdateStatus";

export interface RepairUpdate {
  /** @minLength 1 */
  title?: string;
  location?: string;
  category?: string;
  status?: RepairUpdateStatus;
  priority?: RepairUpdatePriority;
  description?: string;
  /** @minLength 1 */
  reportedBy?: string;
  assignedTo?: string;
  /** @nullable */
  resolvedAt?: Date | null;
  notes?: string;
}